Refactor: Split SchemaItemMoveService into focused move services
