+++ backend/Origam.Architect.Server/Program.cs
@@ -28,6 +28,7 @@ using Origam.Architect.Server.ControlAdapter;
 using Origam.Architect.Server.Interfaces.Services;
 using Origam.Architect.Server.ReturnModels;
 using Origam.Architect.Server.Services;
+using Origam.Architect.Server.Services.Move;
 using Origam.Architect.Server.Services.Wizards;
 using Origam.Architect.Server.Services.Xslt;
 using Origam.Extensions;
@@ -61,6 +62,12 @@ public class Program
         builder.Services.AddTransient<XsltService>();
         builder.Services.AddSingleton<SearchService>();
         builder.Services.AddSingleton<PanelControlFactory>();
+        builder.Services.AddSingleton<MoveNodeResolver>();
+        builder.Services.AddSingleton<MoveRuleEvaluator>();
+        builder.Services.AddSingleton<MoveTargetFinder>();
+        builder.Services.AddSingleton<CrossPackageMoveValidator>();
+        builder.Services.AddSingleton<SchemaItemMover>();
+        builder.Services.AddSingleton<SchemaItemCopier>();
         builder.Services.AddSingleton<SchemaItemMoveService>();
         builder.Services.AddSingleton<PropertyEditorService>();
         builder.Services.AddSingleton<DesignerEditorService>();
